# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "FinToken合约端API文档",
        "title": "FinToken合约端程序",
        "contact": {
            "name": "API Support"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/web/getLocks": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FinLock"
                ],
                "summary": "获取当前钱包拥有的锁列表",
                "parameters": [
                    {
                        "description": "当前钱包信息",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.GetOwnerLocks"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response",
                        "schema": {
                            "$ref": "#/definitions/model.BasicResp"
                        }
                    },
                    "400": {
                        "description": "params error",
                        "schema": {
                            "$ref": "#/definitions/model.BasicResp"
                        }
                    }
                }
            }
        },
...
        "/web/unLock": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "FinLock"
                ],
                "summary": "修改锁状态为pending",
                "parameters": [
                    {
                        "description": "解锁信息",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.ChangeLockStates"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response",
                        "schema": {
                            "$ref": "#/definitions/model.BasicResp"
                        }
                    },
                    "400": {
                        "description": "params error",
                        "schema": {
                            "$ref": "#/definitions/model.BasicResp"
                        }
                    }
                }
            }
        },
        "/web/updateNft": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Wallet"
                ],
                "summary": "更新快照扫下来的nft拥有者的信息",
                "parameters": [
                    {
                        "description": "更新快照扫下来的nft拥有者的信息请求体",
                        "name": "model",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.UpdateNftRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response",
                        "schema": {
                            "$ref": "#/definitions/model.BasicResp"
                        }
                    },
                    "400": {
                        "description": "params error",
                        "schema": {
                            "$ref": "#/definitions/model.BasicResp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "model.BasicResp": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "model.ChangeLockStates": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                }
            }
        },
...
        "model.GetOwnerLocks": {
            "type": "object",
            "properties": {
                "owner": {
                    "type": "string"
                }
            }
        },
...
        "model.UpdateNftRequest": {
            "type": "object",
            "properties": {
                "keys": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "values": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}